+++ FROST-Server.Tests/src/test/java/de/fraunhofer/iosb/ilt/statests/c04batch/BatchResponseJson.java
@@ -66,6 +66,11 @@ public class BatchResponseJson {
         return Objects.equals(this.responses, other.responses);
     }
 
+    @Override
+    public String toString() {
+        return responses.toString();
+    }
+
     public static class ResponsePart {
 
         private String id;